* commit '73fcc88e24638d2efe34469f6a588068ada75638': (29 commits)
  DM-12703 JDBC connection failure in Activiti due to Java security policy - Retry using Activiti engine logic for CannotCreateTransactionException change log level to warn when exception
  Added initial version of the herd.postgres.0.117.0-to-0.118.0.upgrade.sql script
  DM-12703 JDBC connection failure in Activiti due to Java security policy - Retry using Activiti engine logic for CannotCreateTransactionException - add unit test
  DM-12703 JDBC connection failure in Activiti due to Java security policy - Retry using Activiti engine logic for CannotCreateTransactionException reformat only
  DM-12703 JDBC connection failure in Activiti due to Java security policy - Retry using Activiti engine logic for CannotCreateTransactionException
  DM-12409: TestFactory - Herd/Spark - release and generate TSR
  DM-12671: Upgrade JavaMail Dependency - Add exclusion
  changed scala style code
  changed schema version to null
  changed to private
  made changes to DataCatalog
  made changes to DataCatalog
  changes to DC
  added changes to test
  added changes to test
  added test back
  added null to Integer
  added back default usage and file format
  Update Herd pom to 118 SNAPSHOT.
  DM-12660 Handle long values for primary keys in herd entities and model-api objects
  ...

+++ herd-code/herd-dao/src/main/java/org/finra/herd/dao/BusinessObjectFormatDao.java
@@ -66,7 +66,7 @@ public interface BusinessObjectFormatDao extends BaseJpaDao
      *
      * @return the list of business object format ids
      */
-    List<Integer> getBusinessObjectFormatIdsByBusinessObjectDefinition(BusinessObjectDefinitionEntity businessObjectDefinitionEntity);
+    List<Long> getBusinessObjectFormatIdsByBusinessObjectDefinition(BusinessObjectDefinitionEntity businessObjectDefinitionEntity);
 
     /**
      * Gets the maximum available version of the specified business object format.
